requirements.txt
Django==3.1.6
djangorestframework==3.14.0
python-telegram-bot==20.4
Requests==2.31.0
